Training and inference running properly on cluster.
