+++ bin/run_training.py
@@ -1,6 +1,8 @@
 import argparse
-from src.analysis.dataset import ZarrDataset
-from src.analysis.training import run_training
+from analysis.dataset import ZarrDataset
+from analysis.training import run_training
+import pandas as pd
+import torch
 
 
 def parse_args() -> argparse.Namespace:
@@ -19,6 +21,8 @@ def parse_args() -> argparse.Namespace:
                         help='Comma-separated list of training image IDs.')
     parser.add_argument('validation_image_ids', type=str,
                         help='Comma-separated list of validation image IDs.')
+    parser.add_argument('model_name', type=str,
+                        help="Prefix for final model's torchscript file name")
     parser.add_argument('feature_dir_path', type=str,
                         help='Path to the directory containing feature zarr files.')
     parser.add_argument('agg_type', type=str,
@@ -31,13 +35,15 @@ def parse_args() -> argparse.Namespace:
                         help='Name of the target column in the metadata CSV.')
     parser.add_argument('class_order', type=str,
                         help='Comma-separated list of class labels for encoding (e.g., "0,1,2").')
-    parser.add_argument('ordinal', type=str,
+    parser.add_argument('ordinal', type=bool,
                         help='Boolean flag ("True" or "False") to use ordinal encoding.')
 
     # Optional arguments
+    parser.add_argument('--model_out_dir', type=str, default=None,
+                        help="Directory to save the final model's torchscript file to outside of the NextFlow work directory if desired.")
     parser.add_argument('--hidden_dims', type=str, default='128,64',
                         help='Comma-separated list of hidden layer dimensions (e.g., "128,64").')
-    parser.add_argument('--num_epochs', type=int, default=10,
+    parser.add_argument('--num_epochs', type=int, default=1000,
                         help='Number of training epochs.')
     parser.add_argument('--batch_size', type=int, default=32,
                         help='Training batch size.')
@@ -47,7 +53,7 @@ def parse_args() -> argparse.Namespace:
                         help='Dimension of the model output. If not provided, it will be inferred.')
     parser.add_argument('--use_tensorboard', action='store_true',
                         help='Use TensorBoard for logging if this flag is set.')
-    parser.add_argument('--early_stopping_patience', type=int, default=5,
+    parser.add_argument('--early_stopping_patience', type=int, default=10,
                         help='Patience for early stopping.')
     parser.add_argument('--checkpoint_dir', type=str, default='models',
                         help='Directory to save model checkpoints.')
@@ -60,24 +66,29 @@ def parse_args() -> argparse.Namespace:
     return args
 
 
-def main() -> None:
+if __name__ == "__main__":
     args = parse_args()
 
-    training_image_ids = args.training_image_ids.split(',')
-    validation_image_ids = args.validation_image_ids.split(',')
+    training_image_ids = pd.read_csv(args.training_image_ids)["SlideID"].to_list()
+    validation_image_ids = pd.read_csv(args.validation_image_ids)["SlideID"].to_list()
 
     class_order = args.class_order.split(',') if args.class_order else None
     if class_order is not None:
         # Try converting class_order to integers
+        activation_function=torch.nn.Sigmoid()
+        loss_function=torch.nn.BCELoss()
         try:
             class_order = [int(cls) for cls in class_order]
         except ValueError:
             # Keep as strings if not integers
             pass
+    else:
+        activation_function=torch.nn.Identity()
+        loss_function=torch.nn.MSELoss()
 
     hidden_dims = [int(hd) for hd in args.hidden_dims.split(',')]
 
-    ordinal = args.ordinal.lower() == 'true'
+    ordinal = args.ordinal
 
     train_dataset = ZarrDataset(
         image_ids=training_image_ids,
@@ -90,6 +101,7 @@ def main() -> None:
         ordinal=ordinal,
     )
 
+    
     val_dataset = ZarrDataset(
         image_ids=validation_image_ids,
         feature_zarr_dir=args.feature_dir_path,
@@ -128,12 +140,16 @@ def main() -> None:
     }
 
     run_training(
+        model_name=args.model_name,
         train_dataset=train_dataset,
         val_dataset=val_dataset,
         hidden_dims=hidden_dims,
         num_epochs=args.num_epochs,
         batch_size=args.batch_size,
         learning_rate=args.learning_rate,
+        activation_function=activation_function,
+        loss_function=loss_function,
+        model_out_dir=args.model_out_dir,
         output_dim=output_dim,
         early_stopping=early_stopping,
         lr_scheduler=lr_scheduler,
