+++ src/workflows/inference.nf
@@ -1,34 +1,54 @@
 include {CSVCombiner as PatchCSVCombiner} from './modules/CSVHandlingUtils'
 include {CSVCombiner as SlideCSVCombiner} from './modules/CSVHandlingUtils'
 
-process Inference{
+process SlideInference{
     label 'pytorch'
     input:
-        val image_id
+        path image_ids_path
+        path model_path
     output:
-        path 'Patch_Scores_*.csv'
-        path 'Slide_Scores_*.csv'
+        path '*results.csv'
     script:
     """
-    python3 $projectDir/bin/Inference.py $image_id $params.feature_dir_path $params.scan_ds
+    python3 $projectDir/bin/SlideInference.py $image_ids_path $params.feature_dir_path $model_path $params.agg_type_label $params.scan_ds $params.class_order $params.meta_csv $params.results_file_name $params.results_path
     """
 }
 
+process patchInference{
+    label 'pytorch'
+    input:
+        path image_ids_path
+        path model_path
+    output:
+        path "Patch_Scores_*.csv"
+        patth "Mean_Patch_Scores_*.csv"
+    script:
+    """
+    python3 $projectDir/bin/PatchInference.py $image_ids_path $params.feature_dir_path $model_path $params.scan_ds $params.class_order
+    """
+}
 
-
-workflow infer_scores {
+workflow infer_slide_scores {
+    take:
+        image_ids_path
+        model_path
+    main:
+        SlideInference(image_ids_path,model_path)
+}
+workflow infer_patch_scores {
     take:
         image_id
+        model_path
     main:
-        Inference(image_id)
+        PatchInference(image_id,model_path)
         PatchCSVCombiner(Inference.out[0].collect(),params.patch_scores_prefix)
         SlideCSVCombiner(Inference.out[1].collect(),params.slide_scores_prefix)
 }
 
-workflow{
-    Channel.fromPath(params.meta_csv)
-        .splitCSV(header: true, strip: true)
-        .map(row -> row.SlideID)
-        .set(image_id)
-    infer_scores(image_id)
-}
+//workflow{
+//    Channel.fromPath(params.meta_csv)
+//        .splitCSV(header: true, strip: true)
+//        .map(row -> row.SlideID)
+//        .set(image_id)
+//    infer_scores(image_id)
+//}
